2.add-two-numbers/2.add-two-numbers.go

package main

/**
 * <p>You are given two <strong>non-empty</strong> linked lists representing two non-negative integers. The digits are stored in <strong>reverse order</strong>, and each of their nodes contains a single digit. Add the two numbers and return the sum&nbsp;as a linked list.</p>

<p>You may assume the two numbers do not contain any leading zero, except the number 0 itself.</p>

<p>&nbsp;</p>
<p><strong>Example 1:</strong></p>
<img alt="" src="https://assets.leetcode.com/uploads/2020/10/02/addtwonumber1.jpg" style="width: 483px; height: 342px;" />
<pre>
<strong>Input:</strong> l1 = [2,4,3], l2 = [5,6,4]
<strong>Output:</strong> [7,0,8]
<strong>Explanation:</strong> 342 + 465 = 807.
</pre>

<p><strong>Example 2:</strong></p>

<pre>
<strong>Input:</strong> l1 = [0], l2 = [0]
<strong>Output:</strong> [0]
</pre>

<p><strong>Example 3:</strong></p>

<pre>
<strong>Input:</strong> l1 = [9,9,9,9,9,9,9], l2 = [9,9,9,9]
<strong>Output:</strong> [8,9,9,9,0,0,0,1]
</pre>

<p>&nbsp;</p>
<p><strong>Constraints:</strong></p>

<ul>
	<li>The number of nodes in each linked list is in the range <code>[1, 100]</code>.</li>
	<li><code>0 &lt;= Node.val &lt;= 9</code></li>
	<li>It is guaranteed that the list represents a number that does not have leading zeros.</li>
</ul>

**/
/**
 * [2,4,3]
[5,6,4]
**/
/**
 * Definition for singly-linked list.
 * type ListNode struct {
 *     Val int
 *     Next *ListNode
 * }
 */

type ListNode struct {
	Val int
	Next *ListNode
}

func addTwoNumbers(l1 *ListNode, l2 *ListNode) *ListNode {
	lc1 := &l1
	lc2 := &l2
	carry := 0
	var head *ListNode
	current := &head
	for *lc1 != nil || *lc2 != nil || carry != 0 {
		v1 := 0
		if *lc1 != nil {
			v1 = (*lc1).Val
		}

		v2 := 0
		if *lc2 != nil {
			v2 = (*lc2).Val
		}
		*current = &ListNode{
			Val: (v1 + v2 + carry) % 10,
		}
		carry = (v1 + v2 + carry) / 10

		current = &(*current).Next
		if *lc1 != nil {
			lc1 = &(*lc1).Next
		}
		if *lc2 != nil {
			lc2 = &(*lc2).Next
		}
	}

	return head
}
